requirements.txt
gradio
requests
openai
fastapi
uvicorn
gunicorn
azure-identity
azure-keyvault-secrets
dotenv
langchain_openai
langchain_chroma
langchain
tiktoken
bs4
dotenv
langchain-community
azure-search-documents
azure-storage-blob